requirements.txt
langchain
langchain-community
unstructured
pypdf
python-docx
chromadb
python-dotenv